fix: Create the vaults file if it's not exist (#60)
Reviewed-on: https://git.4rs.nl/awiteb/lprs/pulls/60
Co-committed-by: Awiteb <a@4rs.nl>

+++ src/cli/mod.rs
@@ -122,11 +122,16 @@ impl Cli {
             log::info!("Using the given vaults file");
             if let Some(parent) = path.parent() {
                 if parent.to_str() != Some("") && !parent.exists() {
-                    log::info!("Creating the parent vaults file directory");
+                    log::info!(
+                        "Creating the parent vaults file directory: {}",
+                        parent.display()
+                    );
                     fs::create_dir_all(parent)?;
                 }
             }
-            fs::File::create(&path)?;
+            if !path.exists() {
+                fs::File::create(&path)?;
+            }
             path
         } else {
             log::info!("Using the default vaults file");
